messaging module: several changes to the way that messages are sent
Previously, the MessagingService had a reference to all of the gateways and when there was a message to be sent, it passed it to the proper gateway.

Now, MessagingService starts a thread, GatewayManager. GatewayManager polls the database looking for messages with the status "OUTBOX". When messages like this are found, they are passed off to the proper gateway for sending. The only point through which messages can be sent is MessagingService, which just puts the messages in the database with an OUTBOX status.

This approach decouples MessagingGateways from the MessagingService and changes the 'intended audience' of MessagingGateways - previously the gateways were intended to be retrieved and used for sending messages but now they are never directly accessed. This should constitute a more stable and thread-safe approach.