Adding read/unread distinction to messages.
This necessitated a change in how message recipients are represented in the database. Previously, we had only a
join table but now MessageRecipient is its own class. If you
are building and installing this commit, please be advised
that the sqldiff changed without incrementing the version
number.

The UI has not yet been updated to highlight unread messages.